Do not use 64 bit atomics
They do not seem to work on the Rpi 64 bits

+++ src/oscinprocessor.cpp
@@ -41,12 +41,12 @@ void OscInProcessor::prepareOutputs(const vector<string>& outputNames)
 // TODO: during initial testing to measure latency of async calls
 void print_time_stamp(char type);
 
-extern std::atomic<long> g_flush_count;
-extern std::atomic<long> g_midi_send_in_transit;
+extern std::atomic<short> g_flush_count;
+extern std::atomic<short> g_midi_send_in_transit;
 void OscInProcessor::ProcessMessage(long flush_count, const char *c_message, std::size_t size)
 {
     lock_guard<mutex> lock(m_mutex);
-    
+
     // We are processing this message, so reduce number in transit
     g_midi_send_in_transit--;
     // Ignore if this message is old
